scripts/requirements.txt
azure-search-documents==11.4.0
azure-core==1.29.5
azure-identity==1.15.0
openai==1.10.0
supabase==2.3.0
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2
asyncio==3.4.3